homework 1: dto
