+++ bot/src/main/java/edu/java/bot/util/action/AbstractAction.java
@@ -1,7 +1,7 @@
 package edu.java.bot.util.action;
 
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.dto.response.ResponseData;
+import edu.java.bot.util.response.ResponseData;
 import java.util.Optional;
 
 public abstract class AbstractAction implements Action {
