+++ bot/src/main/java/edu/java/bot/util/action/HelpAction.java
@@ -1,8 +1,8 @@
 package edu.java.bot.util.action;
 
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.dto.response.HelpResponse;
-import edu.java.bot.dto.response.ResponseData;
+import edu.java.bot.util.response.HelpResponse;
+import edu.java.bot.util.response.ResponseData;
 import edu.java.bot.util.CommandEnum;
 import java.util.Optional;
 import org.springframework.stereotype.Component;
