+++ bot/src/main/java/edu/java/bot/util/action/UntrackAction.java
@@ -1,8 +1,8 @@
 package edu.java.bot.util.action;
 
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.dto.response.ResponseData;
-import edu.java.bot.dto.response.UntrackResponse;
+import edu.java.bot.util.response.ResponseData;
+import edu.java.bot.util.response.UntrackResponse;
 import edu.java.bot.util.CommandEnum;
 import java.util.List;
 import java.util.Optional;
