+++ bot/src/main/java/edu/java/bot/util/response/HelpResponse.java
@@ -1,4 +1,4 @@
-package edu.java.bot.dto.response;
+package edu.java.bot.util.response;
 
 import com.pengrad.telegrambot.model.Update;
 import edu.java.bot.util.CommandEnum;
