+++ bot/src/main/java/edu/java/bot/util/response/UntrackUrlResponse.java
@@ -1,12 +1,12 @@
-package edu.java.bot.dto.response;
+package edu.java.bot.util.response;
 
 import com.pengrad.telegrambot.model.Update;
 
-public class TrackUrlResponse extends AbstractResponse {
+public class UntrackUrlResponse extends AbstractResponse {
 
     private final String url;
 
-    public TrackUrlResponse(Update update, String url) {
+    public UntrackUrlResponse(Update update, String url) {
         super(update);
 
         this.url = url;
@@ -15,7 +15,7 @@ public class TrackUrlResponse extends AbstractResponse {
     @Override
     @SuppressWarnings("MultipleStringLiterals")
     public ResponseData makeResponse() {
-        String message = "Ссылка " + url + " успешно добавлена для отслеживания.";
+        String message = "Ссылка " + url + " успешно удалена из отслеживания.";
 
         return new ResponseData(createMessageRequest(message), getCommands());
     }
